We the track the position of every field within the query as a
`QueryPosition`. Later it becomes a `Option<QueryPosition>` as fields we
add for requirements don't have any.

We use it to return fields in the right order in the response as that's
part of the GraphQL spec. The problem is that we need to create those
while we're exploding fragments which can result in a huge number of
fields, beyond what `u16` supports. De-duplication can only happen after
as we might need to plan things differently.

As I've changed this, we increased the size of both `ResposneValueId`,
used in the response object path and `FieldShapeRecord`. I could fix the
first. But the second requires really complex logic. Unfortunately, it's
a really sensible struct for the shape generation so it slowed us down
in the worst benchmark case, so we lost 5-7% there. I've improved a bit
the shape generation to handle ids rather than walkers which led to a
sped up of 5% or so compared to `main`.

+++ crates/engine/query-solver/src/solution_space/builder/operation_fields.rs
@@ -52,7 +52,7 @@ struct OperationFieldsIngestor<'schema, 'op, 'builder> {
     builder: &'builder mut QuerySolutionSpaceBuilder<'schema, 'op>,
     // Needs to be a queue to have the right query_position for fields.
     queue: VecDeque<IngestSelectionSet<'op>>,
-    next_query_position: u16,
+    next_query_position: u32,
     // Temporary structures for DFS
     parent_type_conditions: Vec<CompositeTypeId>,
     parent_directive_ids: Vec<ExecutableDirectiveId>,
